inviting_app/requirements.txt
Django==4.2.4
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
